Extract subcomponents into their own files
